@@ -1,47 +1,43 @@
 {
-  "name": "output-file-sync",
-  "version": "2.0.0",
-  "description": "Synchronously write a file and create its ancestor directories if needed",
-  "repository": "shinnn/output-file-sync",
-  "author": "Shinnosuke Watanabe (https://github.com/shinnn)",
-  "scripts": {
-    "pretest": "eslint --fix --format=codeframe index.js test.js",
-    "test": "node --throw-deprecation test.js",
-    "coverage": "node --throw-deprecation node_modules/.bin/istanbul cover test.js"
-  },
-  "license": "ISC",
-  "files": [
-    "index.js"
-  ],
-  "keywords": [
-    "fs",
-    "write",
-    "sync",
-    "synchronous",
-    "output",
-    "create",
-    "file",
-    "mkdir",
-    "mkdirp"
-  ],
-  "dependencies": {
-    "graceful-fs": "^4.1.11",
-    "is-plain-obj": "^1.1.0",
-    "mkdirp": "^0.5.1"
-  },
-  "devDependencies": {
-    "@shinnn/eslint-config-node": "^3.0.0",
-    "eslint": "^3.17.1",
-    "get-mode": "^0.1.0",
-    "istanbul": "^0.4.5",
-    "read-utf8-file": "^1.0.0",
-    "rmfr": "^1.0.2",
-    "tape": "^4.6.3"
-  },
-  "eslintConfig": {
-    "extends": "@shinnn/node",
-    "rules": {
-      "complexity": "off"
-    }
-  }
+	"name": "output-file-sync",
+	"version": "2.0.0",
+	"description": "Synchronously write a file and create its ancestor directories if needed",
+	"repository": "shinnn/output-file-sync",
+	"author": "Shinnosuke Watanabe (https://github.com/shinnn)",
+	"scripts": {
+		"pretest": "eslint --fix --format=codeframe index.js test.js",
+		"test": "nyc --reporter=html --reporter=text node test.js"
+	},
+	"license": "ISC",
+	"files": [
+		"index.js"
+	],
+	"keywords": [
+		"fs",
+		"write",
+		"sync",
+		"synchronous",
+		"output",
+		"create",
+		"file",
+		"mkdir",
+		"mkdirp"
+	],
+	"dependencies": {
+		"graceful-fs": "^4.1.11",
+		"is-plain-obj": "^1.1.0",
+		"mkdirp": "^0.5.1"
+	},
+	"devDependencies": {
+		"@shinnn/eslint-config-node": "^5.0.0",
+		"eslint": "^4.17.0",
+		"get-mode": "^1.0.1",
+		"nyc": "^11.4.1",
+		"read-utf8-file": "^2.0.0",
+		"rmfr": "^2.0.0-3",
+		"tape": "^4.8.0"
+	},
+	"eslintConfig": {
+		"extends": "@shinnn/node"
+	}
 }
